패스북 네이밍 수정
