+++ app/components/PathGroup.css
@@ -157,7 +157,7 @@
   width: 70%;
 }
 
-.bookmark-edit-modal-overlay {
+.pathgroup-edit-modal-overlay {
   position: fixed;
   top: 0;
   left: 0;
@@ -170,7 +170,7 @@
   align-items: center;
 }
 
-.bookmark-edit-modal-background {
+.pathgroup-edit-modal-background {
   background: #ffffff;
   width: 60%;
   max-width: 700px;
@@ -196,7 +196,7 @@
   }
 }
 
-.bookmark-color-modal-background {
+.pathgroup-color-modal-background {
   background: #ffffff;
   width: 40%;
   max-width: 400px; 
@@ -211,14 +211,14 @@
   animation: fadeIn 0.3s ease-in-out;
 }
 
-.bookmark-color-modal-content {
+.pathgroup-color-modal-content {
   display: flex;
   justify-content: center;
   align-items: center;
   padding-top: 2em;
 }
 
-.bookmark-edit-modal-header {
+.pathgroup-edit-modal-header {
   width: 100%;
   display: flex;
   flex-direction: row;
@@ -229,7 +229,7 @@
   font-size: 16px;
 }
 
-.bookmark-edit-done {
+.pathgroup-edit-done {
   position: absolute;
   left: 0.5em;
   top: 0.5em;
@@ -242,44 +242,44 @@
   border-radius: --radius;
   box-shadow: var(--box-shadow-a);
 }
-.bookmark-edit-title {
+.pathgroup-edit-title {
   position: relative;
   font-size: 20px;
   font-weight: 600;
 }
 
-.bookmark-edit-modal-content {
+.pathgroup-edit-modal-content {
   width: 100%;
   height: calc(100% - 3em);
   padding-top: 0.5em;
   overflow-y: auto;
 }
 
-.bookmark-edit-list {
+.pathgroup-edit-list {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   justify-content: center;
   gap: 1em;
 }
 
-.bookmark-folder {
+.pathgroup {
   display: flex;
   flex-direction: column; 
 }
 
-.bookmark-folder-header {
+.pathgroup-header {
   display: flex;
   justify-content: space-between;
   margin-bottom: 1rem;
 }
-.bookmark-folder-header h2 {
+.pathgroup-header h2 {
   align-self: center;
 
   font-size: 1.5rem;
   font-weight: 600;
 }
 
-.add-bookmark-button {
+.add-pathgroup-button {
   all: unset;
   padding: 0.5em 1em;
   background-color: white;
@@ -292,7 +292,7 @@
   box-shadow: var(--box-shadow-a);
 }
 
-.bookmark-add-modal-overlay {
+.pathgroup-add-modal-overlay {
   position: fixed;
   top: 0;
   left: 0;
@@ -304,19 +304,19 @@
   align-items: center;
 }
 
-.bookmark-folder-content {
+.pathgroup-content {
   display: grid;
   grid-template-columns: repeat(3, 1fr);
   justify-content: center;
   gap: 1em;
 }
 
-.choose-bookmark-design {
+.choose-pathgroup-design {
   display: flex;
   flex-direction: column;
   gap: 0.2em;
 }
-.bookmark-design-label {
+.pathgroup-design-label {
   font-size: 0.8em;
   font-weight: 600;
   text-align: left;
@@ -324,20 +324,20 @@
   margin-top: 0.4em;
 }
 
-.preview-bookmark-design-label {
+.preview-pathgroup-design-label {
   font-size: 0.8em;
   font-weight: 600;
   text-align: left;
 }
 
-.preview-bookmark-design {
+.preview-pathgroup-design {
   display: flex;
   flex-direction: row;
   padding: 0.5em;
   align-items: center;
 }
 
-.render-preview-bookmark-design {
+.render-preview-pathgroup-design {
   width: 50%;
   height: 100%;
   display: flex;
@@ -354,7 +354,7 @@
   gap: 0.5em;
 }
 
-.bookmark-title-input {
+.pathgroup-title-input {
   width: 80%;
   padding: 0.3em;
   font-size: 1rem;
@@ -386,7 +386,7 @@
   border-radius: var(--radius);
 }
 
-.bookmark-add-done {
+.pathgroup-add-done {
   width: 75%;
   align-items: center;
   margin-left: 1em;
